implement some Y emps notes 20210509

+++ Yvtu.Core/Entities/MoneyTransfer.cs
@@ -37,7 +37,7 @@ namespace Yvtu.Core.Entities
         public string Note { get; set; }
         public bool Adjusted { get; set; }
         public int AdjustmentNo { get; set; }
-        public int ApiTransaction { get; set; }
+        public long ApiTransaction { get; set; }
         public double FixedFactor { get; set; }
 
     }
@@ -77,7 +77,7 @@ namespace Yvtu.Core.Entities
             public string Note { get; set; }
             public bool Adjusted { get; set; }
             public int AdjustmentNo { get; set; }
-            public int ApiTransaction { get; set; }
+            public long ApiTransaction { get; set; }
             public double FixedFactor { get; set; }
 
         }
