+++ Yvtu.Core/Queries/MoneyTransferQueryDto.cs
@@ -69,6 +69,6 @@ namespace Yvtu.Core.Queries
         public string Note { get; set; }
         public bool Adjusted { get; set; }
         public int AdjustmentNo { get; set; }
-        public int ApiTransaction { get; set; }
+        public long ApiTransaction { get; set; }
     }
 }
