+++ Yvtu.Core/Queries/RechargeQuery.cs
@@ -59,7 +59,7 @@ namespace Yvtu.Core.Queries
         public string RefTransNo { get; set; }
         public DateTime RefTime { get; set; }
         public string DebugInfo { get; set; }
-        public int ApiTransaction { get; set; }
+        public long ApiTransaction { get; set; }
 
     }
 }
