+++ Yvtu.api/Dtos/MoneyTransferDto.cs
@@ -4,7 +4,7 @@
     {
         public string pid { get; set; }
         public int amt { get; set; }
-        public int seq { get; set; }
+        public long seq { get; set; }
         public string note { get; set; }
     }
 }
