+++ Yvtu.api/Dtos/TopupDto.cs
@@ -4,6 +4,6 @@
     {
         public string subsNo { get; set; }
         public int amt { get; set; }
-        public int seq { get; set; }
+        public long seq { get; set; }
     }
 }
